Minor spell-correction, function signature-correction.
ATTENTION!
At the 36th line. I'm not sure that my recommended changes are good.
Accept only if it is.

At the 156th line. The max variable is of type T, but the function
szum() doesn't return type T. So the code is going to be compiled only
if there is implicit conversion from return type of szum function to
type T.

I don't know what to do: replace type T to int (the return type of
szum() function), or something else...